shisha extras price can be float now

+++ client/shisha-rent/src/components/owner/shishasView/shishaUpdate/ShishaUpdate.tsx
@@ -101,7 +101,7 @@ const ShishaUpdate: React.FC = () => {
           ...ps.shishaExtras,
           {
             name: shishaExtrasForm.name,
-            price: parseInt(shishaExtrasForm.price),
+            price: parseFloat(shishaExtrasForm.price),
           },
         ],
       };
